requirements.txt
# Vehicle Damage Assessment - Fraud Detection System
# Python Dependencies

# Core ML and Data Science
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Image Processing
Pillow>=8.3.0
opencv-python>=4.5.0

# Web Application
streamlit>=1.25.0
plotly>=5.0.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
tqdm>=4.62.0
python-dateutil>=2.8.0

# Development and Testing (optional)
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0

# System Requirements
# Python 3.8+ required
# GPU support: CUDA 11.8+ (optional, for faster training)
# RAM: 8GB minimum, 16GB recommended
# Storage: 5GB free space for models and data